修复空指针异常和配置重复bug

+++ src/main/scala/com/mobikok/ssp/data/streaming/util/ModuleTracer.scala
@@ -101,7 +101,7 @@ class ModuleTracer(moduleName: String, config: Config, mixModulesBatchController
 
   def trace (title: String) = {
     val ms = new Date().getTime - lastTraceTime.get()
-    val m = traceBatchUsingTimeFormat.format((100.0*ms/1000/60).asInstanceOf[Int]/100.0)
+    val m = traceBatchUsingTimeFormat.format((100.0*ms/1000/60).asInstanceOf[Long]/100.0)
     if (traceBatchUsingTimeLog.get() == null) {
       startBatch
     }
